pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "masid"
version = "0.1.0"
description = "Multi-Agent System Incentive Design Benchmark"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Inseyab Consulting"},
]
keywords = [
    "multi-agent-systems",
    "llm-agents",
    "mechanism-design",
    "game-theory",
    "benchmark",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "litellm>=1.40.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "pydantic>=2.5",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]
analysis = [
    "pandas>=2.2",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "scipy>=1.13",
    "jupyter>=1.0",
]
crewai = [
    "crewai>=0.80",
    "crewai-tools>=0.14",
]
all = ["masid[dev,analysis,crewai]"]

[project.scripts]
masid = "masid.cli:main"
masid-crewai = "masid_crewai.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require a running LLM endpoint (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
